backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
google-genai==1.73.1
pydantic==2.6.0
